work on making new expression more efficient
This work was somewhat inspired by my work on
MULTIPLER/thermal_resistor.cir, which I ultimately fixed with a
single-line change to the thermal resistor source code.

While I was working on it, I noticed that the old expression library was
called a lot less than the new expression library for the same circuit.

This was because of the "changed" logic that is applied throughout the
device package, so that expressions are only re-evaluated when their
inputs change.  I previously tried to get it this same sort of logic
working in new expression, but messed it up and temporarily
abandoned it.  Today I decided to work on it again.

I mostly have it working, but I haven't yet seen a real performance
improvement with it.  The logical flow thru the code is a bit different
than for the old library.  The old library relied on inputs being
"pushed" to it when the evaluate function was called, so the calling
code could check those inputs and if none changed, then not bother to
call the evaluate function.

The new code handles things via a "pull" operation, so all the logic
needs to sit inside the expression library itself.  99% of the pull
operation is in the funciton  newExpression::getValuesFromGroup, so I
added logic to it to make this work.

As I played around with it, it occured to me that this logic should only
be appplied in the device package usage of the expression library (as
that is how the old library works, essentially).  So, the
"evaluateFunction" functions now have an optional argument, a boolean,
which if set to "true" will instruct the library to attempt this extra
efficiency.

I *think* it is at the point where exact the same tests pass with and
without this efficiency turned on, but for now it is off by default.
I also haven't seen a dramatic speed improvement with it on.  So, it may
be a waste of time, but it was worth trying.   I plan to test it more
before making a final decision on it.

At this point, the number of regression tests that legitimately fail
with the new expression library is pretty small, around 10 I think.
It is a little hard to judge this at the moment b/c I still haven't done
a recent merge with develop, and there has been a lot of code
development there that this branch hasn't benefitted from yet.  For
example, this branch fails the BUG_812_SON test, but I'm sure it will
pass this test as soon as the merge is accomplished.