Fix for PARSER/hspiceMath regression test
I put this off for a while, but it was very easy to fix.  Most Hspice
issues that are relevant to the expression library I have been able to
handle as a "superset".  But, there were (at least) two that couldn't be
handled that way.  The caret "^" operator and the meaning of "log"
simply mean different things in Xyce and Hspice.

In Xyce, the caret means XOR.  In Hspice it means the pow(a,b) operator.

In Xyce, "log" means log10.  In Hspice it means natural log, or ln.

Pete worked on this last year in the old expression library, and to
support this, he created the command line option, -hspice-ext.  Among
other things, this option can set the boolean "useHspiceMath".

In this commit, I've moved the declaration of "useHspiceMath" (as well
as "enableRandomExpression") out of the N_UTL_ExpressionInternals file,
and into the newExpression.C file.  Also, I created a new header for
the "extern" statements for these booleans, N_UTL_HspiceBools.h.

I also have modified the lexer so that the token associated with caret
is more generic.  I have also modified the parser to check the
useHspiceMath boolean, and then allocate the appropriate operators for
"^" and "log" based on its value.