Add optional nodes to BSIM-SOI 4.5.0
This commit adds support for optional p, b, and t nodes to the
BSIM-SOI 4.5.0 (level 70450 MOSFET).

The BSIM-SOI 4.5.0 Verilog-A source code contains "port_connected" use
that is wholly incompatible with Xyce.  I have ifdefed out that usage
and replaced it with the code we use in BSIM-SOI 4.6.1 instead.

As with BSIM-SOI 4.6.1, we exit with fatal errors if the optional
nodes are specified when model parameters indicate that they should
not be used.  The original code specified that they should be
collapsed to ground instead --- but for multiple reasons that turned
out not to work.  When we simply "pretend" that the nodes are
ground (i.e., discard all contributions into those parts of the F and Q
vectors and their matrices), the nodes float and lead to convergence
failures.  And in some cases, the other internal node collapses cause
Xyce/ADMS to fail to process the collapse of the optional nodes to
ground.  Instead, we'll just emit fatal errors when nodes are misused.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39