Regenerate BSIM6 with updated back-end
There was an error in the Xyce/ADMS back-end that I fixed as part of
BSIM-SOI development.  An optional node should not have an internal
node name allocated for it if it is external, nor should it have one
if it's internal and has been collapsed into something else.
Prior to fixing the bug, an optional node would get an internal node
name if it was not connected (i.e. internal), irrespective of whether
it had actually been collapsed.