Various updates, some related to rand/gauss/agauss
There are two types of change in this commit.

The first set of changes is some new book keeping for expressions
containing random number operators such as gauss, agauss, etc.

The second set of changes is a small efficiency improvement.  During
parsing, I have always checked to see if a param invoked by the
expression is invoked more than once.  If so, I only allocate one
object, but then have multiple entries in the tree reference it.

The same thing can be done with other types of nodes.  In particular,
voltage nodes and device currents can be handled this way.  This commit
accopmlishes that.  So now, in an expression such as:

  {20.0*(V(A)**3.0)+7.5*V(A)}

The voltage node "A" will only have a single voltOp allocated for it.
Before this commit, there would have been 2.  In some parts of new
expression, there was enough other book keeping for this not to matter,
but in others not so much.  Also in this case:

  {20.0*(V(A,B)**3.0)+7.5*V(A)}

The code before would have allocated two voltOp nodes for A and one for
B.  With this change it just allocates one of each.