Assure node IDs are in required order
Xyce/ADMS's method of doing optional nodes, which long predates the
ability to add such nodes with $port_connected and used to require
manual patching, works based on the "Node ID" maintained locally in
the instance class.  It *requires* that all external nodes be the
lowest numbered IDs, and internal nodes follow.  It would not work
otherwise.

Unfortunately, the technique I have always used to generate the node
IDs has emitted them based on the order in which they're declared,
taken from ADMS's "node" list.

Unfortunately, some Verilog-A models don't declare their nodes in the
order we require.  For example, BSIM-BULK does this:

  module bsimbulk(d, g, s, b, t);
  inout d, g, s, b, t;
  electrical d, g, s, b, di, si, gi, gm, bi, sbulk, dbulk;
  thermal t;

Nothing about the technique I used to use would *guarantee* that t be
given the highest numbered node of the external nodes, and lower than
all the internals.  In practice, though, it wound up being OK today
when I tried it.  Once upon a time it *wasn't* OK, which accounts for
why in some older Verilog-A models inside Xyce I had to rearrange node
declarations to make sure that the ones that needed to be optional
were declared in just the right order.  At any rate, by simply
iterating over the node list without regard to internal/external, I
was not *assuring* that things were in the right order, I just
happened to luck out that they were.

This commit changes the code that generates the nodeIDs so that all
external nodes are assigned their number first, then the internals.
This *assures* that the node IDs are handed out in the correct order,
instead of being at the mercy of how ADMS stuffs nodes into its node
vector in the ADMS data tree.  Having them always in the right order
makes sure that optional node handling will definitely work out.