Set non-negative ranges on some BSIM-SOI parameters
Some BSIM-SOI parameters will, if negative, cause the model to
misbehave.  Many of them are NOT declared by the BSIM group to have
parameter ranges of any kind, and a user *could* specify them as
negative.  The result would likely be simulation misbehavior with no
indication of why.

This commit adds Verilog-A range specifiers to a number of parameters
that should never be negative.  Some may be zero, others must not be
zero.

Xyce will now emit a warning if a user tries to set these parameters
out of range.  It will not abort just because of the out-of-range
specification.  If the simulation then fails, the user will at least
have had warning that they were doing something wrong, and will be
able to fix the problem.

This commit *only* changes parameters that may not be negative (but
may or may not be zero).  A separate commit will address parameters
that are used in denominators and therefore must never be zero.