Warn if BSIM-SOI parameters in denominators are zero
The BSIM-SOI has a number of parameters that are used directly in
denominators, but were declared without Verilog-A ranges to prevent
them from being zero.

This commit adds range checking to these parameters.

Some of these parameters can be set to zero in other simulators and
perform just fine in DC:  notably, ckappa, which is used only for
computation of overlap capacitance terms, should have no meaning at
DC.  In the BSIM-SOI 3 and the C version of BSIM-SOI 4, these terms
are not even computed at DC.  But in the Verilog-A version of BSIM-SOI
4 they are always computed, and dividing by ckappa=0 makes them be
"nan" (Not a Number).  Because Xyce's DAE formulation is different
from other SPICE-like simulators, the Q matrix is used even at
DC: the presence of these "nan" values in the Q (charge) vector and
consequent nans in the derivative matrix of Q lead to terrible
convergence problems, even at DC.

By adding these warnings, if a user sets the parameters to zero then
they will at least get some notice that they're doing something wrong,
and be pointed at the thing they need to change to fix it.

Xyce will still fail simulations that other simulators breeze through
at DC, but those other simulators would also fail with the same model
card in transient or AC.  The correct fix is not to set parameters to
zero when those parameters are being used in denominators.

The BSIM-SOI model itself should have been checking that.  It's got
many other sanity checks on parameters and computed values.  It's just
such a complicated model that stuff like this could easily be
overlooked, especially in a model that dates back as far as this one,
with its SPICE C legacy.