Give up trying to support TNODEOUT in 5- and 6-node use
The BSIM-SOI goes through some effort to support 5- and 6- node
instances with TNODEOUT=1, trying to make the last node be
temperature.

Unfortunately, while the Verilog-A source does have some code to treat
a final node as temperature in this case, in other places they make
explicit assumptions that the 5th node is the external body tie and
the 6th the internal body node.  These assumptions are completely
incompatible with the attempt to make the "last" node temperature when
it isn't the 7th node.

I have tried my best to hack on the model to make this use case work,
and have instead dined on a 7 course meal of nothing with a side order
of bupkis.  I am giving up.  Doing it right would mean reworking the
BSIM-SOI source code too much to still call it a faithful
implementation.

If a user attempts:
   M1 d g s e t model TNODEOUT=1
or
   M1 d g s e p t model TNODEOUT=1

Xyce will exit with a fatal error saying it cannot implement this
use.  Only the 7-node version can work, with the explicitly given
internal body node as the 6th node:

   M1 d g s e p b t model TNODEOUT=1

However, some use cases of *this* instance won't work, either, because
they try to collapse the p and b nodes together with some choices of
model parameter.  Xyce cannot collapse external nodes together.  These
use cases, too, will exit with error.

When not specified, the temperature node is internal when SHMOD=1, and
Xyce provides a means of accessing that node on print lines.  If all
that is needed is to output the temperature rise of the device, it can
be obtained through the N() operator with the appropriate internal
node name (generally, the instance name with the suffix "_t").

This does mean you can't use the external temperature node of the
device to create a thermal coupling network between devices, unless
one is using the full 7-node version of the device.

I wish it were otherwise.  Unfortunately, as the model is written in
Verilog-A, the technique chosen to implement the optional temperature
node with TNODEOUT is not compatible with Xyce.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39