partial implementation for random operators
I have been working on connecting random operators such as AGAUSS, etc
to the UQ capabilities such as SAMPLING, EMBEDDEDSAMPLING and PCE.

This is an interim commit on this topic.  This is several days of work,
and I wanted to describe it before I forgot what I did.

To make things like AGAUSS, etc work with sampling, it is necessary to
be able to gather together the relevant information and put it into an
organized data structure that the UQ analysis classes can use to setup
samples.

In the old, non-expression way that SAMPLING worked, a list of
parameters would be specified on the .SAMPLING line.  During the parsing
of that line, those parameters are loaded into a vector of "sampling
params", or N_ANP_SweepParam objects.  At various stages of the
analysis, this vector is looped over and the information in it is used
to create the sample points or update them.

This commit mainly accomplishes the task of sending a query from
N_ANP_Sampling to the device manager, requesting a vector of
N_ANP_SweepParam objects, that have been populated with information from
various random expression operators such as AGAUSS.

The device manager is the only place in the code that has enough
infomration to gather that information.

This gathering procedure seems to work, at least in serial.  I belive
that it will not (yet) work in parallel.  Also, I havne't yet added code
to make the "setParam" aspect of this work yet.  So, the vector of
parameters is set up, and the UQ code can properly set up all the sample
points.  But right now if it tries to loop over them and set their
values, that won't work yet.  That is what I am working on today.

Note, with the old expression library, the random values for operators
such as AGAUSS were computed directly there, in the operator.  This was
reasonable to do because at the time they were not being sampled and
thus other parts of the code didn't need to know anything about them.

For various reasons, the UQ parts of the code need to know a list of the
random parameters, and they need to compute all the sample points before
the analysis starts.  For example, if you want to apply a covariance
matrix, you'll need all the samples in one place.  Similarly, if you
want to use PCE via quadrature, you need all the information in one
place to compute the quadrature points.

Another small thing that I had to modify with this commit is the
handling of immutable values in the Util::Param object.    Now that
expressions with random operators can be changed (as we go from one
sample step to the next) it was necessary to add logic to ensure
that any expression depending on random operators is considered
non-constant.

When you call "Util::Param::getImmutableValue<double>" (or int, or
any supported type), the code will check if the parameter is of type
Util::EXPR, and if it is, it will then check if it is a constant
expression.  The expression needs to be constant b/c the parameter
will get casted by this call to be of Util::DBLE type.  (and thus can
never be treated as an expression afterwards).  If getImmutableValue
is called on a non-constant Util::EXPR parameter, the code throws a
fatal error.

I have struggled a little bit with the logic of when an expression is
constant and when it isn't, and in this part of the code I was getting
it wrong.   Fortunately, most parts of the code (i.e. the parser) that
call "getImmutableValue" check this before making the call. (otherwise
we'd get fatal errors frequently)

All of this came up b/c in the device manager, resistor values are
checked to see if they are below a threshold, and if they are the
Resistor object is replaced with a Resistor3 object.  When it does this
check, it uses the "getImmutableValue" call to do it.  I wasn't checking
for random operators in either place (device mgr or Util::Param) so the
call went thru just fine, and the parameter magically got changed from
Util::EXPR to Util::DBLE.  Hence, my attempt to gather up parameters
depending on random operators would then sometimes fail.

So, to make my new code work properly (and also b/c it was the right
thbing to do) ehis commit updates the logic in Util::Param to do a
more detailed check of the const-ness of an expression, including
random operators.