Document BSIM-SOI optional nodes
Now that we have a limited support for the optional nodes of BSIM-SOI
4.x, I am documenting that support and its limitations in the
reference guide and 7.2 release notes.

In doing so, I have modified some of the documentation of BSIM-SOI
3.2 (level=10).  Our MOSFET documentation has BSIM-SOI in a fairly
prominent place, because BSIM-SOI 3.2 was a big deal in early
development of Xyce --- that documentation had a lot of statements
about the BSIM-SOI that were specific to 3.2 and would easily be
mistaken by a user as referring to the newer BSIM-SOI models as well.
It now describes the BSIM-SOI as being 3 different level numbers, and
where something is true only for the level=10, we say so

This is primarily about TNODEOUT, which cannot currently be supported
in 4.6.1 and 4.5 because of the way they handle the P and B nodes, and
the half-correct way they try to kludge TNODEOUT into the Verilog-A
version --- this sort of multi-function optional node handling is much
easier in the SPICE-specific C code where it originated than it is in
Verilog-A, but is a terrifying nightmare in Xyce no matter how it's
written, as evidenced by the nightmarish code in N_DEV_MOSFET_B3SOI.C to
handle it.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39