Bring generic switch documentation up to date
Many years ago, Xyce had three independent switch devices:  the
original voltage- and current- controlled switches (S and W) and a
"generic" switch controllable by an expression (SW) (which was added
somewhere around 2001 as part of an import of ChileSPICE features).

Almost as many years ago, in 2004, the original voltage- and
current-controlled switch code was disposed of, and in their stead
some code was added to the input parser to convert switch lines in
those formats into equivalent use of the generic switch.  At that
time, the "SW" device disappeared and became the "S" device --- the
difference between the voltage-controlled S and the generic-controlled
S divined from context (if it uses CONTROL and only has 2 nodes, it's
generic, if it uses the four-node version it's voltage-controlled and
trivially converted to generic with CONTROL={V(node3,node4)}).

When that happened, the documentation was never updated.  It continued
to document the "SW" device as if it were real.

Of course, the examples used in that documentation still worked (as do
all of our tests of the generic device), because "SW" starts with "S".
So a user following the documentation exactly would not exactly be led
astray.  But they might unnecessarily have an extra "W" in their
device name that served no real purpose.

This commit brings the documentation in line with reality after 16
years of being wrong.  It was not noticed that it was wrong until an
open source user asked about it.

Closes https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/22