more steps toward sampling/expression integration
This commit is a checkpoint commit.  This code compiles, and doesn't
break anything (relative to what was already broken a few days ago).

For my previous commit(s) on this work, of integrating the UQ capabilities
like SAMPLING to the new expression library operators like AGAUSS was all about
gathering all the relevant operators so that the SAMPLING analysis class could
know what they were, and to bundle them up in a data structure similar to the
one already used by that class to set up samples.

This commit is more about the other half of the process, where once the
sampling loop has started, at each iteration of that loop the sample
values have to be updated.

For the old sampling specification, that was all in terms of parameters
and using the DeviceMgr::setParam function call (and all the related
function such as the free function "setParameter" in the
N_DEV_DeviceMgr.C file) to set the value of that parameter.

As it was written, the DeviceMgr::setParam function was not adequate to
handle random operators inside of type=Util::EXPR parameters.  It works
OK if the type=Util::EXPR parameter is a simple expression that ONLY
includes the random operator and nothing more.  So, for example:

.global_param eric={agauss(1.0, 0.01, 1.0)}

This works with the existing setParam structure, b/c the value of the
operator agauss is exactly equal to the value of the param.

But of course, part of the point of using AGAUSS inside an expression is
to allow for it to be part of a larger, more complex expression.  So,
for example, this would need to work as well:

.global_param par1={sqrt(pi)}
.global_param eric={par1*agauss(1.0, 0.01, 1.0) + rand() + time}

In the above example, it would be completely incorrect to update the
parameter "eric" to the value of either the agauss or rand operators, as
they are just a part of the expression.  So, the existing
"DeviceMgr::setParam" structure wasn't enough.

I chose to leave the existing setParam structure alone, rather than
hacking it (and potentially breaking it).  So, I've developed about 80%
of a parallel structure for setting the values on randomized operators
in expressions.  It follows the same patterns, roughly, but a little
more information has to be communicated; it isn't enough to say which
parameter we are modifying, it is also necessary to say which of the
(possibly more than one) random operators in the expression we are
modifying.  This is mostly set up now, except for one thing.  The one
thing that is NOT set up is the actual setting of the operator value
in the expression.

For the most part, I've used a "pull" procedure (via expression groups)
rather than a "push" procedure (like setParam) to set values on
expressions.  I plan to keep doing this, but the above pattern I just
described sort of goes in the opposite direction.  So, this seemed like
a good place to stop for now.

Next, I'm going to go back to the new expression library and work on the
"pull" methodology for random operators, and create some unit tests for
it.