Implement first unit tests of random ops
This commit does two things.

(1) in the random op classes such as AGAUSS, there is now an
"useHspiceMath" conditional, where if it is "true", the operator just
returns the mean of the distribution.  If it is false, it returns zero,
which is WRONG.   That use case will be refined later.

(2) adds first unit tests of random ops such as AGAUSS.  Like the
refinements I added to the ops themselves, these tests are necessary but
not sufficient.  They will pass if useHspiceMath is true and fail if it
is false.  So, they only test the use case of returning the mean, which
is what Hspice does for the non-sampling use case.