Minor cleanup to update param functions
The "setParameter" free function (which is uniquely called from
DeviceMgr::setParam had kind of a weird set of function arguments.  This
commit does a minor revision.

The new free function "setParameterRandomExpressionTerms" (as it was
created initially as a copy of setParameter, then revised) had a very
similar set of function arguments.  So, it got the same minor revision.
The revision was probably more important for this function.

Why are these free functions?  b/c Dave Baur made them that way.  Why
did he do it? probably so that they could also be used by
VECTOR-COMPOSITE related parameters.  That is my guess.  He never
finished that refactor, however so it isn't clear.

So, what was the revision in this commit?

There is a object that is owned by the device mgr, called "globals_".
This object is of type "Global", which is defined in the
N_DEV_SolverState.h header.  It is currently defined as:

struct Globals
{
  GlobalParameterMap global_params;
  std::vector<Util::Expression> global_expressions;
  std::vector<std::string>global_exp_names;
};

ie, an object containing 3 different STL containers related to global
parameters.

The "setParameter" free function only needed the first of these (the
global_params map), and so it just took that as a function argument.
But I found this confusing, in part b/c it changed the name for
no obvious reason.

The new "setParameterRandomExpressionTerms" free function needs all 3 of
them.  And, will probably need more than that (i.e. "Globals" will
probabaly get expanded soon).  As I am still iterating on this, it made
much more sense for the argument to setParameterRandomExpressionTerms
simply be "globals" rather than have a bunch of separate args for each
part of globals.  Especially if those parts will be changing as I
work thru this.

So, the above change was more important for the new function
("setParameterRandomExpressionTerms"), but it felt like the right
thing to do for the old one ("setParameter") as well.