Improve collectAssignedVariables logic
Ever since 2008, when I was working on the original Xyce/ADMS and
trying to get it good enough to process even VBIC 1.2, I have relied
on a template called "collectAssignedVariables" to figure out what
variables to declare --- the idea was that you'd only declare
variables that are actually assigned to.  It works by recursively
drilling down the ADMS data tree and finding all variables assigned to
at every level of the tree, and making sure to propagate that fact
upward through the tree, so that at top level we know which variables
to declare.

The reason for using such a method is something that is lost in
development history.  The commit message of the commit that introduced
it does not explain my reasoning, but it was necessary to get even
simple toy models compiled, according to that message.  It is such a
strange thing to do (rather than just declare all variables that
exist) that I am pretty sure that I cribbed the technique off of the
handful of existing open-source ADMS back-ends that I had available to
me back in those days.  I also recall having read commentary from
developers of other ADMS back-ends that they were doing it that way,
too, but if they were, they aren't anymore.

It does appear that this is important to the handling of
initial_instance and initial_model blocks, but not to deal with
GLOBAL_INSTANCE or GLOBAL_MODEL scoped variables (because those don't
get declared locally, they get declared as member variables of the
instance or model classes).

At any rate, while working on MEXTRAM in 2014, I found that there were
some cases where block-local variables and module-scoped variables of
the same name were confusing this method and causing multiple
declarations.  I "fixed" the problem by refusing to propagate a
variable upward if the enclosing context already had such a variable
listed as assigned to elsewhere.

This was wrong, as it causes a block-local variable that is assigned
to in a block that appears *before* an assignment that assigns to a
module-scoped variable of the same name to be the only one present in
the "@assignedVars" list at the module level.

In fact, the correct approach would have been only to propagate an
assigned variable upward if it is in fact scoped at a level *higher*
than the current block.  A block-local variable should not be pushed
up to a higher level if the current level is the block where it's
declared.

This commit changes the logic so that "assignedVars" are only pushed
upward by a block-level item if they are not declared in this block.

This change correctly keeps block-local information local to a block,
and eliminates the issue of block-local and module-scoped variables of
the same name not being declared properly.

This issue is essential for getting HiSIM_HV done right.  Fixing it
corrects the problem that led to its detection: a local Mu_res
variable declared both at module and local scope, but used at both
scopes --- the local scope happened to be earlier in the code than the
module scope, and so was not being detected as an "assignedVar" due to
the use of the name as the condition for pushing upward, and therefore
not declared in both places.

As a result of this change, several existing Verilog-A models now get
generated with fewer local variables declared in
updateIntermediateVars or the instance- or model-class processParams
functions.  As it turned out, the declarations that go away were of
variables that were never used, and should never have been
declared (they were block-local, and are also declared at block level
in the C++ emitted --- only the local versions are actually used).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/40