Try to fix Xyce.sty to match new SAND style
The SAND style master branch removed the "[table]" option from its
xcolor usage, and this causes Xyce docs to fail now because we *ALSO*
require xcolor and were doing it with the "[table]" option.  I am trying to
see if removing that option returns us to being able to build, and if
so, if it doesn't break our documents in some visual way.