Revert "Try to fix Xyce.sty to match new SAND style"
This reverts commit 8815cf9deb44dd1b9a99028d111295a3cb61c62a.

Removing [table] from the xcolor package does NOT fix our documents.

We are apparently *using* xcolor's [table] option, and so the fact
that the SAND style author has removed it (thereby clashing with our
usage) makes it incompatible.

I am going to have to work with the SAND style maintainers to see if
he is willing to add [table] back.