Try to fix Xyce.sty to work with new SAND style
They removed the "[table]" option from the SAND style's xcolor
include, which made it incompatible with Xyce.sty.

Removing "[table]" from our usage broke our docs in a different way,
but that's because "[table]" causes LaTeX to add the "colortbl"
package under the hood, and that package provides the \rowcolor that
we need.

I'm trying to fix this by removing "[table]" and adding a
RequirePackage{colortbl} to see if that does the trick.