Improve Handling of WHEN Time for FIND-WHEN and WHEN
This commit fixes the handling of the WHEN time for the FIND-WHEN
and WHEN measures, for AC, DC and NOISE measure modes, if either
FROM or TO are specified.  The FIND-WHEN and WHEN measures are now
only satisfied if the  interpolated WHEN time is within the
specified measurement window.  This commit also further ensures
that the AT time, for a FIND-AT measure, must be within the FROM-TO
window for a successful measure.

As part of this change, the functionality for updating and evaluating
the RISE, FALL and CROSS conditions have been moved to member functions
in the FindWhenBase class.  These new functions are much simpler now,
since the "when time" is a known crossing, and "absolute mode" and
the RFC_LEVEL qualifier are no longer supported for WHEN-based
measures.  Those two new functions will likely be moved to the Base
class, as support for the RISE, FALL and CROSS qualifiers is added
for the DERIV-WHEN measure for AC, DC and NOISE measure modes.

The new isInvalidTimeFreqWindow() and isInvalidDCsweepWindow()
Base class functions provide some computational efficiency for
WHEN-based measures. They also preserve the convention that a WHEN-
or AT-based measure is "initialized" if there is at least one
simulation step within a valid FROM-TO window.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304