Remove RFC_LEVEL for DERIV-WHEN, FIND-WHEN and WHEN
This commit removes support for the RFC_LEVEL qualifier for the
DERIV-WHEN, FIND-WHEN and WHEN measure types.  That qualifier is
not supported in HSPICE for any measure type.  It was added to
Xyce to improve the performance of rise, fall and cross counting
for MAX, MIN and PP measures (where there is no WHEN clause on the
.MEASURE) and for use with the Xyce-specific FRAC_MAX qualifier
for TRIG-TARG measures.  It is not particularly useful for WHEN-based
measures.

For backwards compatibility, a warning message is emitted if the
RFC_LEVEL qualifier is used on a DERIV-WHEN, FIND-WHEN or WHEN
measure line.  The requested RFC_LEVEL value is then ignored during
subsequent processing for that measure.

SON Bug 1314
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1314