Make Xyce/ADMS not emit unnecessary second derivatives
When I implemented ddx, I made Xyce/ADMS emit second derivatives when
it thinks it needs to.

Second derivatives are needed if a quantity that is obtained through a
ddx() call is used in a static or dynamic (excluding noise)
contribution, because we need the derivative for the jacobian.

Unfortunately, the logic for determining when this was necessary was
not correct, and in part was not correct because the original implicit ADMS
rules from which we have created "our" adms.implicit.xml set the
"insource" flag of a variable if it is used in any way in
contributions, including noise contributions.  This made it difficult
to write the *correct* condition.

Our logic was to emit second derivatives for any quantity if that
quantity was "insource" or if it included "ddxprobes" --- that is, if
any subsequent variable impacted by this one was going to be
differentiated by ddx.

That is the wrong logic --- it doesn't matter if *we're* insource, it
matters if the variable that gets computed through ddx() is in a
non-noise source.  But detecting *THAT* is really hard, and so far I
haven't been able to work through the logic needed to propagate that
information through the ADMS data tree.

Furthermore, the implicit rules incorrectly propagate ddxprobes.  In
code  like this:
   CSBI = ddx(QSI,v(body))
   CSB = CSBI + ddx(QG,v(sbulk))

should make QG have a ddxprobe, and any variable that goes into
computing QG gets it propagated backward as well.

BUT: the implicit rules were not doing this propagation to QG --- they
were propagating ddxprobes into every variable in the expression on
the RHS, unnecessarily marking CSBI as having a ddxprobe.  Since CSBI
is itself a variable with ddx value, it appeared that second
derivatives of QSI were needed (because it mistakenly was saying we
needed to differentiate CSBI by the ddxprobes).

The result of these issues was that Xyce/ADMS was generating huge
amounts of completely unnecessary second derivative code, NONE of it
actually being used.  This is the reason that BSIM-SOI takes geologic
timescales to compile.

So to minimize the unnecessary output, I've made a number of changes
in this commit:
  - Change the implicit rules to set insource only for non-noise
    contributions.  Xyce only uses "insource" to figure out if it needs
    derivatives, and the noise implementation in Xyce does NOT need
    them.
  - Change the handling of ddx in the implicit rules so that ddxprobes
    are ONLY pushed into the variables present in the first argument
    of the ddx() call, not to every variable in the entire
    expression *containing* a ddx() call.
  - Wrap all code in xyceBasicTemplates_nosac.xml that generates
    second derivatives (and declarations of the variables needed,
    and any use of those variables) in a conditional that skips them
    unless a global variable called "doSecondDerivs" is set to "yes"
  - Add code to xyceImplementationFile_nosac.xml and
    xyceHeaderFile_nosac.xml to set "doSecondDerivs" to "yes" only if
    there exists any variable in the module with both "insource='yes'"
    and "derivate='yes'".  "derivate='yes'" means that the variable is
    set to an expression containing ddx.  Thus, we only get second
    derivatives if any such variable is used in non-noise contributions.

Note that in the case of using a "derivate='yes'" variable in
non-noise contributions, we're currently "all-or-nothing" --- any use
of such a variable anywhere results in all second derivative code
generation being enabled, not just generation of *needed*
second-derivatives.  At this time, the implicit rules do not push
enough information through the tree to make it easy to detect.  The
result is that such usage will restore *all* the bloat we're getting
rid of with this commit in models that *don't* use things that way.

Use of ddx() variables in contributions is already something that is
discouraged in best-practices documents.  We have never encountered a
model that does it, except in noise contributions.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/42