Fix for coilgun ddx test
One of the remaining failing tests with the new expression library has
been the COILGUN2/eric_coilgun_4-coil_ddx.cir test.

I initially thought that the reason it was failing was because I did
something wrong with the ddx operator.  But it turns out that there was
no problem with it.

I then thought that maybe the use of ddx with .func was the problem, but
that was fine as well.

Finally, I figured out that the problem did not have anything to do with
ddxOp at all.  The real problem was with the tableOp class.  I had not
created any unit tests that relied on the tableOp::dx function.  As
such, I didn't catch that it was set up incorrectly.    At the time
that I wrote it, I was so enamored with the fact that the new expression
library could handle tables with operators for each entry in the table,
that I focused on derivatives of table entries, rather than derivatives
of the input.

This function sets up dx to compute (if necessary) derivatives of the
input.  I initially tried to set it up using the "evalDeriv" function in
the linear interpolator class.  But, while this did produce derivatives
that seemed roughly correct, it was insufficient to get the
eric_coilgun_4-coil_ddx.cir test to pass.  So, I gave up on that
approach and implemented the equivalent of what was done in the old
expression library.  Once I did that, the test passed.

This takes the number of failing tests down to 14 (excluding tests that
fail due to divergence between the develop branch and this one)

This commit also includes a new unit test related to this issue (use of
ddx with .func and tables).